requirements.txt
python-dotenv==1.0.0
selenium==4.15.2
python-dateutil==2.8.2
dateparser==1.2.0
beautifulsoup4==4.12.2
pymongo==4.6.1
fastapi==0.104.1
uvicorn==0.24.0.post1
python-jose==3.3.0
python-multipart==0.0.6